Replace our eslint configuration with Airbnb's
I think Airbnb's javascript rules are becoming an industry standard at
this point; and they are certainly more extensive than our existing
eslint configuration at least.

I went through and checked that all rules we had enabled are configured
the same way in Airbnb's ruleset -- they are. So this should be a bit
more painful until we fix the possibly-new styling lints, but we can
address that on a file-by-file basis.