Allow seeds script to be run against `localhost` db
There is a safeguard in the seeds script to not nuke a
production/staging db by accident. This is cool, but I want to use it
locally in development.

To facilitate that, a check for if the db is localhost should suffice!

This also moves the CCENV check to use the credentials exported value
instead of pulling it from the environment again.