Add Users.findAllByClientIds method
This method accepts an array of client IDs ("clid") and returns an array
of the associated user objects (as referenced by the foreign key "cm").

This can be useful when displaying a table of clients to include the
full names of the case managers. The method accepts an array so as to
prevent N+1 queries.