Fix display of case manager name on client index view
The case manager name had been showing as "undefined undefined" due to
the database normalization that happens in
`BaseModel._getMultiResponse` and the fact that the organization-level
view of this page doesn't join against those columns anyway.

By running a subsequent query serially to return the associated users
for the visible clients, and then indexing those users by the "cm" /
"cmid" foreign key / primary key, we can display the user objects
without hacking them into the model's schema.

If we end up doing this often, we may wish to switch to an ORM that lets
us define these associations up front and then traverse them on-demand.

Fixes #315