Add "baseUrl" configuration option for homepage
The baseUrl option, e.g. 'https://secure.clientcomm.org', is the
deployment's canonical URL. Currently this is just used in two places:
generating the twilio callback URLs and the homepage's login button.

I updated the configuration value to include the scheme as well since it
helps when rendering a link.

Fixes #331