Pass `deploy_base_url` to Chef as attribute
This configuration variable should be passed to Chef as an attribute so
that it can be used in configuration files. (It probably makes sense to
pass all configuration values this way instead of creating a config file
with terraform.)

To do this, a JSON structure is output to /etc/chef/attributes.json on
each server and then chef is configured to load attributes from that
file in the /etc/chef/solo.rb config file.

This also moves most of the chef logic into a bash function that is
called for each web server.