Allow run-chef.sh script to accept a single IP to run on
For testing purposes, as well as deploy purposes, it will be helpful for
this script to accept a single IP to run on.

This will allow the deploy process to run on each node in serial, so we
can catch deploy problems before they are rolled out to all the servers.